ENG-99: Move lineup lock and earliest game start time logic to custom hook
